requirements.txt
# OpenRouter + OpenAI SDK
openai>=1.0.0

# Document processing
python-docx>=0.8.11

# Data handling
pandas>=1.5.0
openpyxl>=3.0.0
rapidfuzz>=3.0.0

# Type validation
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Optional: For async support
aiohttp>=3.8.0
